Fixed data type conversion warning with MSVC.

+++ src/OffscreenContextWGL.cpp
@@ -282,7 +282,7 @@ void checkWglFeatures(sgl::vk::Device* device) {
          */
         //dxgiAdapterDesc1.VendorId;
         ComPtr<IDXGIOutput> out;
-        for (size_t odx = 0; dxgiAdapter1->EnumOutputs(odx, &out) == S_OK; odx++) {
+        for (size_t odx = 0; dxgiAdapter1->EnumOutputs((UINT)odx, &out) == S_OK; odx++) {
             DXGI_OUTPUT_DESC desc2;
             out->GetDesc(&desc2);
             displayName = sgl::wideStringArrayToStdString(desc2.DeviceName);
